start moving browse logic into own component (WIP)
